# docs/openapi-minecraft.yaml
openapi: 3.1.0
info:
  title: Minecraft MCP API
  description: REST API exposed by the Minecraft Fabric mod on port 7070. Wraps world manipulation, build task management, and prefab placement.
  version: 0.0.1
servers:
  - url: http://localhost:7070
    description: Local Minecraft server
...
paths:
...
  /api/world/structure/place:
    parameters:
      - $ref: "#/components/parameters/AreaLockHeader"
    post:
      tags: [structures]
      summary: Place an NBT structure (multipart upload)
      operationId: placeNbtStructure
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/NbtStructurePlaceRequest"
      responses:
        "409":
          $ref: "#/components/responses/AreaLockConflict"
        "200":
          description: Structure placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NbtStructurePlaceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"

  # ── Effects ──────────────────────────────────────────────────────────────────
  /api/world/effects/rain-fire:
    parameters:
      - $ref: "#/components/parameters/AreaLockHeader"
    post:
      tags: [effects]
      summary: Rain fire over a circular area
      operationId: rainFire
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RainFireRequest"
      responses:
        "409":
          $ref: "#/components/responses/AreaLockConflict"
        "200":
          description: Fire placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RainFireResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"

  # ── Builds ───────────────────────────────────────────────────────────────────
...
components:
  parameters:
    AreaLockHeader:
      name: X-Area-Lock-Id
      in: header
      required: false
      description: Optional reservation token. Writes must fit inside its world and cuboid. Unknown or expired tokens fail; absence only permits writes outside reserved areas. Execute/replay carries the token into each task and stops on a lock failure.
      schema:
        type: string
...
  responses:
...
    AreaLockConflict:
      description: Conflicting reservation, write outside its reservation, or unknown/expired token. No blocks are written by the rejected operation. Queued task failures are reported in build status instead.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AreaLockError"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    AreaBounds:
      type: object
      description: Inclusive block bounds; each minimum must not exceed its maximum.
      required: [min_x, min_y, min_z, max_x, max_y, max_z]
      properties:
        min_x:
          type: integer
          format: int32
        min_y:
          type: integer
          format: int32
        min_z:
          type: integer
          format: int32
        max_x:
          type: integer
          format: int32
        max_y:
          type: integer
          format: int32
        max_z:
          type: integer
          format: int32
    AreaLockMetadata:
      type: object
      required: [world, bounds, label, expires_at, idle_timeout_seconds]
      properties:
        world:
          type: string
        bounds:
          $ref: "#/components/schemas/AreaBounds"
        label:
          type: string
        expires_at:
          type: string
          format: date-time
        idle_timeout_seconds:
          type: integer
          default: 900
...
    AreaLockError:
      type: object
      required: [success, code, error]
      properties:
        success:
          type: boolean
          const: false
        code:
          type: string
          enum: [area_locked, outside_area_lock, invalid_area_lock]
        error:
          type: string
        reservation:
          $ref: "#/components/schemas/AreaLockMetadata"
        requested_world:
          type: string
        requested_bounds:
          type: object
          description: Inclusive operation bounds; empty for a no-op with a mismatched world.

    # ── Shared primitives ──────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
    World:
      type: string
      default: minecraft:overworld
      example: minecraft:overworld
    Position3D:
      type: object
      properties:
        x:
          type: number
        y:
          type: number
        z:
          type: number
      required: [x, y, z]
...
    BlockRotation:
      type: string
      enum: [NONE, CLOCKWISE_90, CLOCKWISE_180, COUNTERCLOCKWISE_90]
      default: NONE
...
    NbtStructurePlaceRequest:
      type: object
      properties:
        nbt_file:
          type: string
          format: binary
          description: NBT structure file
        world:
          type: string
          default: minecraft:overworld
        x:
          type: integer
          default: 0
        y:
          type: integer
          default: 64
        z:
          type: integer
          default: 0
        rotation:
          $ref: "#/components/schemas/BlockRotation"
        include_entities:
          type: string
          enum: ["true", "false"]
          default: "true"
        replace_blocks:
          type: string
          enum: ["true", "false"]
          default: "true"
      required: [nbt_file]
    NbtStructurePlaceResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        filename:
          type: string
        position:
          $ref: "#/components/schemas/Position3D"
        world:
          type: string
        structure_size:
          $ref: "#/components/schemas/Position3D"
        rotation:
          type: string
        include_entities:
          type: boolean
        replace_blocks:
          type: boolean

    # ── Effects ───────────────────────────────────────────────────────────────
    RainFireRequest:
      type: object
      properties:
        world:
          $ref: "#/components/schemas/World"
        x:
          type: integer
          description: Center X
        z:
          type: integer
          description: Center Z
        radius:
          type: integer
          minimum: 1
          maximum: 56
        density:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
        seed:
          type: integer
          format: int64
      required: [x, z, radius, density]
    RainFireResponse:
      type: object
      properties:
        success:
          type: boolean
        world:
          type: string
        fires_placed:
          type: integer
        columns_considered:
          type: integer
        center:
          type: object
          properties:
            x:
              type: integer
            z:
              type: integer
        radius:
          type: integer
        density:
          type: number

    # ── Builds ────────────────────────────────────────────────────────────────
...